pyproject.toml
[project]
name = "code-pair-partner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "fastapi-mcp>=0.2.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.1",
    "crewai[tools]>=0.114.0",
    "mcp-run>=0.4.4",
    "crewai-tools[mcp]>=0.40.1",
    "mcpadapt>=0.1.3",
    "pygithub>=2.6.1",
]

[tool.uv]
dev-dependencies = [
    "pytest-coverage>=0.0",
    "pytest>=8.3.5",
    "rust>=1.3.1",
]

[tool.crewai]
type = "crew"
